refactor : minify code and delete home.tsx
