+++ apps/designer/components/Layout.tsx
@@ -1,15 +1,23 @@
 import Navbar from "./Navbar";
 import React from "react";
 import { ScriptProps } from "next/script";
-import HomePage from "./HomePage";
 import LoadingBar from "./elements/LoadingBar";
+import { useState } from "react";
+import Welcome from "./Welcome";
+import NewDesign from "./NewDesign";
 const Layout: React.FC<ScriptProps> = ({ children }) => {
   // TODO: improve this type later
+  const [state, setState] = useState(false);
+
+  const handleClick = () => {
+    setState(true);
+  };
   return (
     <>
       <Navbar />
       {children}
-      <HomePage />
+      {state === false && <NewDesign handleClick={handleClick} />}
+      {state === true && <Welcome />}
       <LoadingBar />
     </>
   );
