+++ apps/designer/pages/_app.tsx
@@ -8,7 +8,7 @@ const MyApp: NextComponentType<AppContext, AppInitialProps, AppLayoutProps> = ({
   pageProps,
 }: AppLayoutProps) => {
   return (
-    <div className="h-full bg-pink-100 ">
+    <div className="h-full bg-pink-200 ">
       <Layout>
         <Component {...pageProps} />
       </Layout>
